backend/requirements.txt
# Backend Dependencies
Django==5.0.2
python-dotenv==1.0.0
mysqlclient==2.2.4  # Para la conexión con MySQL
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pytest==8.0.0  # Para pruebas 
pytest-django==4.8.0  # Para pruebas de Django
